refactor: move BackupPlan zipping files to a separate directory; remove unneeded SaveBackupPlanCommand; change files to use file scoped namespaces
