+++ src/BackupHelper.Core/BackupZipping/IBackupPlanZipper.cs
@@ -1,0 +1,6 @@
+namespace BackupHelper.Core.BackupZipping;
+
+public interface IBackupPlanZipper
+{
+    void CreateZipFile(BackupPlan plan, string outputPath);
+}
