+++ src/BackupHelper.Core/Features/CreateBackupCommand.cs
@@ -1,39 +1,38 @@
-﻿using BackupHelper.Core.FileZipping;
+﻿using BackupHelper.Core.BackupZipping;
 using MediatR;
 using Microsoft.Extensions.Logging;
 
-namespace BackupHelper.Core.Features
+namespace BackupHelper.Core.Features;
+
+public record CreateBackupCommand(BackupPlan BackupPlan, params string[] BackupFilePaths) : IRequest<CreateBackupCommandResult>;
+public record CreateBackupCommandResult;
+
+public class CreateBackupCommandHandler : IRequestHandler<CreateBackupCommand, CreateBackupCommandResult>
 {
-    public record CreateBackupCommand(BackupPlan BackupPlan, params string[] BackupFilePaths) : IRequest<CreateBackupCommandResult>;
-    public record CreateBackupCommandResult;
+    private readonly ILogger<CreateBackupCommandHandler> _logger;
+    private readonly IBackupPlanZipper _backupPlanZipper;
 
-    public class CreateBackupCommandHandler : IRequestHandler<CreateBackupCommand, CreateBackupCommandResult>
+    public CreateBackupCommandHandler(ILogger<CreateBackupCommandHandler> logger, IBackupPlanZipper backupPlanZipper)
     {
-        private readonly ILogger<CreateBackupCommandHandler> _logger;
-        private readonly IBackupPlanZipper _backupPlanZipper;
+        _logger = logger;
+        _backupPlanZipper = backupPlanZipper;
+    }
 
-        public CreateBackupCommandHandler(ILogger<CreateBackupCommandHandler> logger, IBackupPlanZipper backupPlanZipper)
+    public Task<CreateBackupCommandResult> Handle(CreateBackupCommand request, CancellationToken cancellationToken)
+    {
+        foreach (var backupFilePath in request.BackupFilePaths)
         {
-            _logger = logger;
-            _backupPlanZipper = backupPlanZipper;
-        }
-
-        public Task<CreateBackupCommandResult> Handle(CreateBackupCommand request, CancellationToken cancellationToken)
-        {
-            foreach (var backupFilePath in request.BackupFilePaths)
+            try
             {
-                try
-                {
-                    _backupPlanZipper.CreateZipFile(request.BackupPlan, backupFilePath);
-                    _logger.LogInformation("Backup created at {BackupFilePath}", backupFilePath);
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogError(ex, "Failed to create backup at {BackupFilePath}", backupFilePath);
-                }
+                _backupPlanZipper.CreateZipFile(request.BackupPlan, backupFilePath);
+                _logger.LogInformation("Backup created at {BackupFilePath}", backupFilePath);
             }
-
-            return Task.FromResult(new CreateBackupCommandResult());
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "Failed to create backup at {BackupFilePath}", backupFilePath);
+            }
         }
+
+        return Task.FromResult(new CreateBackupCommandResult());
     }
 }
