--- src/BackupHelper.Core/Features/SaveBackupPlanCommand.cs
@@ -1,19 +1,0 @@
-﻿using BackupHelper.Core.FileZipping;
-using MediatR;
-
-namespace BackupHelper.Core.Features
-{
-    public record SaveBackupPlanCommand(BackupPlan BackupPlan, string ConfigurationSavePath) : IRequest<SaveBackupPlanCommandResult>;
-
-    public record SaveBackupPlanCommandResult;
-
-    public class SaveBackupPlanCommandHandler : IRequestHandler<SaveBackupPlanCommand, SaveBackupPlanCommandResult>
-    {
-        public Task<SaveBackupPlanCommandResult> Handle(SaveBackupPlanCommand request, CancellationToken cancellationToken)
-        {
-            request.BackupPlan.ToJsonFile(request.ConfigurationSavePath);
-
-            return Task.FromResult(new SaveBackupPlanCommandResult());
-        }
-    }
-}
