--- src/BackupHelper.Core/FileZipping/IBackupPlanZipper.cs
@@ -1,7 +1,0 @@
-namespace BackupHelper.Core.FileZipping
-{
-    public interface IBackupPlanZipper
-    {
-        void CreateZipFile(BackupPlan plan, string outputPath);
-    }
-}
